+++ lib/philiprehberger/env_validator/schema.rb
@@ -16,8 +16,8 @@ module Philiprehberger
       # @param required [Boolean] whether the variable is required
       # @param default [String, nil] default value if not set
       # @return [void]
-      def string(name, required: false, default: nil)
-        @definitions[name.to_s] = { type: :string, required: required, default: default }
+      def string(name, required: false, default: nil, choices: nil)
+        @definitions[name.to_s] = { type: :string, required: required, default: default, choices: choices }
       end
 
       # Define an integer variable.
@@ -25,9 +25,10 @@ module Philiprehberger
       # @param name [Symbol, String] the ENV variable name
       # @param required [Boolean] whether the variable is required
       # @param default [Integer, nil] default value if not set
+      # @param choices [Array, nil] allowed values
       # @return [void]
-      def integer(name, required: false, default: nil)
-        @definitions[name.to_s] = { type: :integer, required: required, default: default }
+      def integer(name, required: false, default: nil, choices: nil)
+        @definitions[name.to_s] = { type: :integer, required: required, default: default, choices: choices }
       end
 
       # Define a float variable.
@@ -35,9 +36,10 @@ module Philiprehberger
       # @param name [Symbol, String] the ENV variable name
       # @param required [Boolean] whether the variable is required
       # @param default [Float, nil] default value if not set
+      # @param choices [Array, nil] allowed values
       # @return [void]
-      def float(name, required: false, default: nil)
-        @definitions[name.to_s] = { type: :float, required: required, default: default }
+      def float(name, required: false, default: nil, choices: nil)
+        @definitions[name.to_s] = { type: :float, required: required, default: default, choices: choices }
       end
 
       # Define a boolean variable.
@@ -45,9 +47,10 @@ module Philiprehberger
       # @param name [Symbol, String] the ENV variable name
       # @param required [Boolean] whether the variable is required
       # @param default [Boolean, nil] default value if not set
+      # @param choices [Array, nil] allowed values
       # @return [void]
-      def boolean(name, required: false, default: nil)
-        @definitions[name.to_s] = { type: :boolean, required: required, default: default }
+      def boolean(name, required: false, default: nil, choices: nil)
+        @definitions[name.to_s] = { type: :boolean, required: required, default: default, choices: choices }
       end
     end
   end